+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/cockroachdb/cockroach-operator/compare/v2.10.0...master)
+# [Unreleased](https://github.com/cockroachdb/cockroach-operator/compare/v2.11.0...master)
+
+# [v2.11.0](https://github.com/cockroachdb/cockroach-operator/compare/v2.10.0...v2.11.0)
+
+* Bug fix to return correct error message if value passed for `whenUnsatisfiable` under `topologySpreadConstraints` field is invalid.
+* Add openshift release process for cockroach operator
 
 # [v2.10.0](https://github.com/cockroachdb/cockroach-operator/compare/v2.9.0...v2.10.0)
 
@@ -31,6 +36,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Delete the CancelLoop function, fixing a cluster status update bug
 * Correctly detect failed version checker Pods
 * retry cluster status updates, reducing test flakes
+
+## Changed
+* Update validation webhook to reject changes to cluster spec's AdditionalLabels field
 
 # [v2.7.0](https://github.com/cockroachdb/cockroach-operator/compare/v2.6.0...v2.7.0)